@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class BlockingServiceApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(BlockingServiceApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(BlockingServiceApplication.class, args);
+	}
 
 }
